feat: 角色 ID 保留區段 + system/developer 系統保留角色後台不可見
AclRoleSeeder 指定固定 ID（全域 1-10、admin.* 從 51 起），新增 system 角色；
RoleController 以 SYSTEM_ROLES 常數統一排除 system/developer 的列表、編輯、刪除；
同步 super_admin 自動指派所有啟用權限邏輯、權限 seeder 改用 access/modify/delete 命名。

+++ database/migrations/0001_01_03_000001_create_acl_tables.php
@@ -47,7 +47,7 @@ return new class extends Migration
             $table->string('guard_name');
 
             // 自訂欄位
-            $table->unsignedInteger('sort_order')->default(0);
+            $table->unsignedSmallInteger('sort_order')->default(0);
             $table->boolean('is_active')->default(true);
 
             $table->timestamps();
